Separate JS code to modules
