+++ src/view/controls-container.js
@@ -1,0 +1,1 @@
+export const createControlsContainer = () => `<div class="trip-main__trip-controls  trip-controls"></div>`;
