+++ src/view/events-container.js
@@ -1,0 +1,1 @@
+export const createTripEventsContainer = () => `<section class="trip-events"></section>`;
