+++ src/view/events-list.js
@@ -1,0 +1,1 @@
+export const createTripEventsList = () => `<ul class="trip-events__list"></ul>`;
