+++ src/view/new-event-button.js
@@ -1,0 +1,1 @@
+export const createNewEventButton = () => `<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>`;
