+++ src/view/trip-days-list.js
@@ -1,0 +1,1 @@
+export const createTripDaysList = () => `<ul class="trip-days"></ul>`;
