+++ src/view/trip-info.js
@@ -1,0 +1,1 @@
+export const createTripInfo = () => `<section class="trip-main__trip-info  trip-info"></section>`;
